requirements.txt
python-telegram-bot==20.7
yt-dlp==2023.12.30
aiohttp==3.9.1
aiofiles==23.2.1
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
instaloader==4.10.3
tweepy==4.14.0